requirements.txt
# Bronze Level Dependencies
watchdog>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# Silver Level Dependencies
# Gmail Integration
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# MCP Servers (Flask)
flask>=3.0.0
requests>=2.31.0

# Scheduling
pywin32>=306; sys_platform == 'win32'
